# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/v1/hot-topic/{community}/solution": {
            "post": {
                "security": [
                    {
                        "Internal": []
                    }
                ],
                "description": "add topic solution",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "HotTopic"
                ],
                "summary": "ToReview",
                "parameters": [
                    {
                        "type": "string",
                        "description": "lowercase community name, like openubmc, cann",
                        "name": "community",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "body",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controller.reqToAddSolution"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/controller.ResponseData"
                        }
                    }
                }
            }
        },
        "/v1/hot-topic/{community}/to-review": {
            "post": {
                "security": [
                    {
                        "Internal": []
                    }
                ],
                "description": "upload topics to review",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "TopicReview"
                ],
                "summary": "Create",
                "parameters": [
                    {
                        "type": "string",
                        "description": "lowercase community name, like openubmc, cann",
                        "name": "community",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "body",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controller.reqToReview"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/controller.ResponseData"
                        }
                    }
                }
            }
        },
        "/v1/topic-review/{community}": {
            "get": {
                "security": [
                    {
                        "Internal": []
                    }
                ],
                "description": "get topic review info",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "TopicReview"
                ],
                "summary": "Get",
                "parameters": [
                    {
                        "type": "string",
                        "description": "lowercase community name, like openubmc, cann",
                        "name": "community",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/app.TopicsToReviewDTO"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "Internal": []
                    }
                ],
                "description": "update the selected topics",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "TopicReview"
                ],
                "summary": "Update",
                "parameters": [
                    {
                        "type": "string",
                        "description": "lowercase community name, like openubmc, cann",
                        "name": "community",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "body",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controller.reqToUpdateSelected"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Accepted",
                        "schema": {
                            "$ref": "#/definitions/controller.ResponseData"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "app.DiscussionSourceInfo": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "source_closed": {
                    "type": "boolean"
                },
                "source_id": {
                    "type": "string"
                },
                "source_type": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                }
            }
        },
        "app.OptionalTopic": {
            "type": "object",
            "properties": {
                "discussion": {
                    "type": "array",
                    "items": {
                        "type": "array",
                        "items": {
                            "$ref": "#/definitions/app.DiscussionSourceInfo"
                        }
                    }
                },
                "summary": {
                    "type": "string"
                }
            }
        },
        "app.TopicsToReviewDTO": {
            "type": "object",
            "properties": {
                "cadidates": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "array",
                        "items": {
                            "$ref": "#/definitions/domain.TopicToReview"
                        }
                    }
                },
                "selected": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.TopicToReview"
                    }
                }
            }
        },
        "controller.ResponseData": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "data": {},
                "msg": {
                    "type": "string"
                }
            }
        },
        "controller.reqToAddSolution": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/app.OptionalTopic"
                    }
                }
            }
        },
        "controller.reqToReview": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/app.OptionalTopic"
                    }
                }
            }
        },
        "controller.reqToUpdateSelected": {
            "type": "object",
            "properties": {
                "selected": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.TopicToReview"
                    }
                }
            }
        },
        "domain.DiscussionSourceToReview": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "imported_at": {
                    "type": "string"
                },
                "source_closed": {
                    "type": "boolean"
                },
                "source_id": {
                    "type": "string"
                },
                "source_type": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                }
            }
        },
        "domain.TopicToReview": {
            "type": "object",
            "properties": {
                "dss": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.DiscussionSourceToReview"
                    }
                },
                "order": {
                    "type": "integer"
                },
                "resolved": {
                    "type": "boolean"
                },
                "title": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "Internal": {
            "description": "Type \"Internal\" followed by a space and internal token.",
            "type": "apiKey",
            "name": "TOKEN",
            "in": "header"
        }
    }
}